// codegen/EndpointSettings.ts
/* tslint:disable */
/* eslint-disable */
/*
 * Autogenerated by @creditkarma/thrift-typescript v3.7.6
 * DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
*/
import * as thrift from "thrift";
export interface IEndpointSettingsArgs {
    joinMode?: number;
}
export class EndpointSettings {
    public joinMode?: number;
    constructor(args?: IEndpointSettingsArgs) {
        if (args != null && args.joinMode != null) {
            this.joinMode = args.joinMode;
        }
    }
    public write(output: thrift.TProtocol): void {
        output.writeStructBegin("EndpointSettings");
        if (this.joinMode != null) {
            output.writeFieldBegin("joinMode", thrift.Thrift.Type.I32, 1);
            output.writeI32(this.joinMode);
            output.writeFieldEnd();
        }
        output.writeFieldStop();
        output.writeStructEnd();
        return;
    }
    public static read(input: thrift.TProtocol): EndpointSettings {
        input.readStructBegin();
        let _args: any = {};
        while (true) {
            const ret: thrift.TField = input.readFieldBegin();
            const fieldType: thrift.Thrift.Type = ret.ftype;
            const fieldId: number = ret.fid;
            if (fieldType === thrift.Thrift.Type.STOP) {
                break;
            }
            switch (fieldId) {
                case 1:
                    if (fieldType === thrift.Thrift.Type.I32) {
                        const value_1: number = input.readI32();
                        _args.joinMode = value_1;
                    }
                    else {
                        input.skip(fieldType);
                    }
                    break;
                default: {
                    input.skip(fieldType);
                }
            }
            input.readFieldEnd();
        }
        input.readStructEnd();
        return new EndpointSettings(_args);
    }
}