// codegen/HangupRequest.ts
/* tslint:disable */
/* eslint-disable */
/*
 * Autogenerated by @creditkarma/thrift-typescript v3.7.6
 * DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
*/
import * as thrift from "thrift";
export interface IHangupRequestArgs {
    reason?: number;
    detailedReasonString?: string;
}
export class HangupRequest {
    public reason?: number;
    public detailedReasonString?: string;
    constructor(args?: IHangupRequestArgs) {
        if (args != null && args.reason != null) {
            this.reason = args.reason;
        }
        if (args != null && args.detailedReasonString != null) {
            this.detailedReasonString = args.detailedReasonString;
        }
    }
    public write(output: thrift.TProtocol): void {
        output.writeStructBegin("HangupRequest");
        if (this.reason != null) {
            output.writeFieldBegin("reason", thrift.Thrift.Type.I32, 1);
            output.writeI32(this.reason);
            output.writeFieldEnd();
        }
        if (this.detailedReasonString != null) {
            output.writeFieldBegin("detailedReasonString", thrift.Thrift.Type.STRING, 2);
            output.writeString(this.detailedReasonString);
            output.writeFieldEnd();
        }
        output.writeFieldStop();
        output.writeStructEnd();
        return;
    }
    public static read(input: thrift.TProtocol): HangupRequest {
        input.readStructBegin();
        let _args: any = {};
        while (true) {
            const ret: thrift.TField = input.readFieldBegin();
            const fieldType: thrift.Thrift.Type = ret.ftype;
            const fieldId: number = ret.fid;
            if (fieldType === thrift.Thrift.Type.STOP) {
                break;
            }
            switch (fieldId) {
                case 1:
                    if (fieldType === thrift.Thrift.Type.I32) {
                        const value_1: number = input.readI32();
                        _args.reason = value_1;
                    }
                    else {
                        input.skip(fieldType);
                    }
                    break;
                case 2:
                    if (fieldType === thrift.Thrift.Type.STRING) {
                        const value_2: string = input.readString();
                        _args.detailedReasonString = value_2;
                    }
                    else {
                        input.skip(fieldType);
                    }
                    break;
                default: {
                    input.skip(fieldType);
                }
            }
            input.readFieldEnd();
        }
        input.readStructEnd();
        return new HangupRequest(_args);
    }
}